requirements/base.txt
fastapi==0.103.2
pydantic==2.4.2
email-validator==2.0.0.post2
pydantic_settings==2.0.3
uvicorn==0.23.2
pytz==2023.3.post1


aiokafka==0.8.1

loguru==0.7.2

asyncpg==0.28.0
alembic-1.12.0
sqlalchemy==2.0.22

pytest==7.4.2
sqlalchemy_utils==0.41.1